+++ bank-feeds/CodatBankFeeds/Models/Shared/DataConnectionStatus.cs
@@ -60,4 +60,5 @@ namespace Codat.BankFeeds.Models.Shared
             throw new Exception($"Unknown value {value} for enum DataConnectionStatus");
         }
     }
+
 }