+++ bank-feeds/CodatBankFeeds/Models/Shared/DataType.cs
@@ -138,4 +138,5 @@ namespace Codat.BankFeeds.Models.Shared
             throw new Exception($"Unknown value {value} for enum DataType");
         }
     }
+
 }