+++ bank-feeds/CodatBankFeeds/Models/Shared/PushChangeType.cs
@@ -62,4 +62,5 @@ namespace Codat.BankFeeds.Models.Shared
             throw new Exception($"Unknown value {value} for enum PushChangeType");
         }
     }
+
 }